Fix for #2500 - installer now complains if you miss information. (other than display name - that's not required).
Also made changes regarding #2650 - which is a won't fix bug.

Explanation:

realpath() is partially used for getting propper paths to files, sure. So that part of it could be worked around (the usage in class e107 I'm talking
about here). *but* realpath is also used as a major security check. For example, it's used to prevent request.php allowing the download of any file in
the filesystem - for example - e107_config.php.

In the future it may even be used more with the threat from arbitary file inclusion being what it is, both local and remote. The security benefits of
using a function like realpath() are too great to ignore.

Thus the decision is that we do not support hosts that disable the realpath() function whatsoever. If your host does this, chances are that the rest of
the PHP install is so badly castrated and the servers are so over-sold that the host isn't worth using anyways.

One such host is the free host Awardspace. Just goes to prove that you get what you pay for.

The changes I've made will report if the function is disabled right from the start in the installer.

svn path=/trunk/e107_0.7/; revision=7470