Make it possible for non admins to download more files
There seemed to be a bug that only admins can download files located in the system import folder. It is possible that this was some kind of security mechanism at one point although calling this function would require PHP access to modify what you are attempting to download anyway, so if there is something sensitive in that files location it would be vulnerable anyway.