Permissive field validation in db_verify
Even if it's a new installation of e107, `db_verify` will be too strict in
validating database field defaults that are equivalent to each other.

This can happen on some MySQL or MariaDB servers that correct the field
definitions, which were not strictly correct in the table structures built into
e107 in the first place.

This commit introduces db_verify::validateFieldPermissive(), a
backwards-compatible way to validate database field defaults.  It works
regardless of whether the MySQL server stores a corrected representation of the
schema and does not require fixing the schema in any existing table structure
files.

Fixes: #2998