Allow spaces in passwords system-wide by default
Some e107 users and admins may use whitespace in their passwords, but the
client-side HTML validation currently prohibited them by default.

The server accepts passwords with whitespace if this client-side validation is
disabled, so there does not appear to be a technical reason for this
restriction.

This commit allows whitespaces in passwords by default.

Fixes: #3006