Deactivated "stash" lock in GitPreparer
The "stash" lock was used to shove .gitignore'd files under the rug
so that they would not interfere with a pure copy of the app.

Vendor files may be ignored in the app, so for performance, the
"stash" lock has been deactivated. Vendor files no longer need to
be downloaded each time the test runs.

The "commit" lock now includes all ignored files so that tests
are run with the files as they are.