PHP bug 66773 workaround in Hybridauth
In PHP 7.0.12 and earlier, a class could not be loaded with the same
name as another class in a different namespace.  This commit patches the
Hybridauth dependency to avoid PHP bug 66773.  Unfortunately, the fix
will be rolled back any time Hybridauth is updated or overwritten.

PHP bug link: https://bugs.php.net/bug.php?id=66773