#3867: Accurate relative path calculations in e107::set_paths()
e107 historically conflated e_BASE with a URI path and a local file
system path. e_BASE seems to have been redefined later to mean a
relative path to the e107 root from the calling script, and e_HTTP was
introduced to resolve URIs from the web browser. e_ROOT is the absolute
path represented by e_BASE.

Because of legacy usage of e_BASE depending on it being a relative path,
e_BASE must remain as a relative path, but how the path was determined
was incorrectly implemented.

This commit fixes multiple issues with e107::set_paths():

* e_BASE is now a relative path calculated reliably by a helper function
* If ./e107_handlers/e107_class.php is in a sensible place and
  ./class2.php can be found one directory up, this path will be used as
  the traversal target.
* If ./e107_handlers/e107_class.php is at an exotic path,
  debug_backtrace() will be inspected to find "class2.php", and a
  relative path will be made to the dirname() of that class2.php
* In CLI mode, chdir() is now called to set the working directory to the
  e107 root. This is to maintain relative path consistency. Previously,
  the absolute path would be stored in e_BASE, which may lead to
  inconsistent behavior.
* e_HTTP is now resolved from $_SERVER['SCRIPT_NAME'] instead of
  $_SERVER['PHP_SELF'] because arbitrary strings and slashes can be
  added to the end of $_SERVER['PHP_SELF'] and could lead to e_HTTP
  storing URIs that descended too far.

Fixes: #3867