adding custom class to be able style just PM link
Now it has clean bootstrap markup. If you have a higher menu, the position of PM icon is wrong (it is correct, the height of the icon is different than the avatar or text itself), but you need a class to be able to fix this position without changing other stuff.