Fix: news category breadcrumbs
The breadcrumbs for news categories were previously pointing to the
category page but using the id of the news story instead of the id of
the category.

Previously, if you had two news stories with ids 1 and 2, and one
category with id 1, then news story 1 would link to its category,
but news story 2 would link to a 404 page.